# lib/api-zod/src/generated/types/evaluationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewSession } from "./interviewSession";
import type { QuestionResult } from "./questionResult";

export interface EvaluationResult {
  session: InterviewSession;
  questionResults: QuestionResult[];
}