# lib/api-zod/src/generated/types/interviewAnswer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewAnswer {
  id: number;
  sessionId: number;
  questionId: number;
  userAnswer: string;
  idealAnswer?: string | null;
  score?: number | null;
  feedback?: string | null;
  strengths?: string | null;
  weaknesses?: string | null;
}