# lib/api-zod/src/generated/types/interviewQuestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewQuestion {
  id: number;
  sessionId: number;
  questionNumber: number;
  question: string;
  topic: string;
  subTopic?: string | null;
  questionType: string;
  similarQuestion?: string | null;
  companyTags?: string | null;
}