# lib/api-zod/src/generated/types/interviewSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewSession {
  id: number;
  userId: string;
  interviewType: string;
  difficulty: string;
  jobRole: string;
  company?: string | null;
  skills?: string | null;
  status: string;
  totalScore?: number | null;
  accuracyPercent?: number | null;
  timeTakenSeconds?: number | null;
  strengths?: string[] | null;
  weaknesses?: string[] | null;
  suggestions?: string[] | null;
  createdAt: Date;
  completedAt?: Date | null;
}