# lib/api-zod/src/generated/types/interviewSessionWithQuestions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewQuestion } from "./interviewQuestion";
import type { InterviewSession } from "./interviewSession";

export interface InterviewSessionWithQuestions {
  session: InterviewSession;
  questions: InterviewQuestion[];
}