# lib/api-zod/src/generated/types/questionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuestionResult {
  questionId: number;
  question: string;
  topic: string;
  questionType: string;
  similarQuestion?: string | null;
  companyTags?: string | null;
  userAnswer: string;
  idealAnswer: string;
  score: number;
  feedback: string;
  strengths: string;
  weaknesses: string;
}