# lib/api-zod/src/generated/types/startInterviewBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StartInterviewBodyDifficulty } from "./startInterviewBodyDifficulty";
import type { StartInterviewBodyInterviewType } from "./startInterviewBodyInterviewType";

export interface StartInterviewBody {
  interviewType: StartInterviewBodyInterviewType;
  difficulty: StartInterviewBodyDifficulty;
  jobRole: string;
  company?: string;
  skills?: string;
}